requirements-dev.txt
# ─────────────────────────────────────────────────────────────────────────────
# Development & test dependencies — NOT bundled into Lambda package
# Install with: pip install -r requirements-dev.txt
# ─────────────────────────────────────────────────────────────────────────────
-r requirements.txt

# Testing
pytest>=8.0
pytest-cov>=5.0
pytest-mock>=3.12
moto[cloudwatch,logs,lambda,dynamodb]>=5.0

# Linting & formatting
ruff>=0.4

# Type checking
mypy>=1.9
boto3-stubs[cloudwatch,logs,bedrock-runtime,lambda]>=1.34

# Environment management
python-dotenv>=1.0
